+ Add utils pickDate && add date-range-input e2e test
